pyproject.toml
[project]
name = "aura-plugins"
version = "0.9.9"
description = "Parallel agent launcher for Aura multi-agent workflows"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "temporalio>=1.9.0",
    "pyyaml>=6.0.3",
    "aiosqlite>=0.22.1",
    "jinja2>=3.1.6",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "pymysql>=1.1"]

[tool.setuptools.packages.find]
where = ["scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["scripts", "tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
